Fix hash(channel) overflow at INT_MIN in Postgres shard routing
hash(channel) can be exactly -2147483648 (INT_MIN) for a rare channel name, and
abs() overflows there.

Go (read-replica routing): the routing hash returned int32 and the caller took
abs() before the modulo; abs(math.MinInt32) stays negative, so getReadPool
computed a negative replica index and panicked when read replicas are configured.
Make hashtext return uint32 - an unsigned hash is always non-negative, so the
modulo is taken directly with no abs and no overflow edge case. The hash only
selects an interchangeable read-replica pool (nothing persisted, not sent to
Postgres), so changing it is non-breaking.

SQL (shard assignment): abs(hashtext(...)) raised "integer out of range"
(hashtext is int4), aborting the publish/remove/expire function for that channel
permanently. Cast to bigint before abs so the value fits; magnitude is unchanged
for every other channel, so no shard is remapped.

The Go change fixes every deployment (runtime code). The SQL change only edits
schema.sql, so it applies to fresh installs immediately and to existing databases
on their next schema-version bump (the upgrade path re-applies schema.sql's
functions) - no dedicated migration, matching the ~1/2^32 rarity.

+++ internal/pgmapbroker/internal/sql/schema.sql
@@ -193,7 +193,7 @@ BEGIN
     END IF;
 
     -- Calculate shard_id from channel hash
-    v_shard_id := abs(hashtext(p_channel)) % p_num_shards;
+    v_shard_id := abs(hashtext(p_channel)::bigint) % p_num_shards;
 
     -- 0. Per-shard serialization lock (lock order: shard → meta → state)
     PERFORM 1 FROM __PREFIX__shard_lock WHERE shard_id = v_shard_id FOR UPDATE;
@@ -472,7 +472,7 @@ BEGIN
     END IF;
 
     -- Calculate shard_id from channel hash
-    v_shard_id := abs(hashtext(p_channel)) % p_num_shards;
+    v_shard_id := abs(hashtext(p_channel)::bigint) % p_num_shards;
 
     -- 0. Per-shard serialization lock (lock order: shard → meta → state)
     PERFORM 1 FROM __PREFIX__shard_lock WHERE shard_id = v_shard_id FOR UPDATE;
@@ -638,7 +638,7 @@ BEGIN
     LOOP
         -- 0. Calculate shard_id and acquire per-shard serialization lock
         --    (lock order: shard → meta → state).
-        v_shard_id := abs(hashtext(v_channel)) % p_num_shards;
+        v_shard_id := abs(hashtext(v_channel)::bigint) % p_num_shards;
         PERFORM 1 FROM __PREFIX__shard_lock WHERE shard_id = v_shard_id FOR UPDATE;
 
         -- 1. Lock meta (same order as publish: shard → meta → state).
